+++ src/level/__init__.py
@@ -1,6 +1,8 @@
 from .abstractions import SpritesManager
-from .concretes import LevelManager, ObstaclesManager, ScoreObserver,CoinObserver
+from .concretes import LevelManager
 from .interfaces import ILevelManager
+from .observers import CoinObserver, ScoreObserver
+from .sprites import ObstaclesManager
 
 __all__ = [
     "ObstaclesManager",
@@ -8,5 +10,5 @@ __all__ = [
     "ILevelManager",
     "ScoreObserver",
     "SpritesManager",
-    "CoinObserver"
+    "CoinObserver",
 ]
