--- src/level/concretes/entity_observers/__init__.py
@@ -1,4 +1,0 @@
-from .score_observer import ScoreObserver
-from .coin_observer import CoinObserver
-
-__all__ = ["ScoreObserver","CoinObserver"]
