+++ src/level/concretes/level_manager.py
@@ -2,10 +2,10 @@ from pygame import Surface
 
 from src.entities import Hero, IElementObserver
 from src.enums import HeroType, Level, World
-from src.utils import PLAYER_LIVES, Camera
+from src.utils import HERO_LIVES, Camera
 
 from ..interfaces import ILevelManager
-from .sprites_managers import ObstaclesManager
+from ..sprites import ObstaclesManager
 
 
 class LevelManager(ILevelManager):
@@ -22,7 +22,7 @@ class LevelManager(ILevelManager):
         score_observer: IElementObserver,
         coin_observer: IElementObserver,
         camera: Camera,
-        lives: int = PLAYER_LIVES,
+        lives: int = HERO_LIVES,
     ) -> None:
         self.__hero = hero
         self.__hero_type = hero_type
@@ -90,7 +90,7 @@ class LevelManager(ILevelManager):
     def get_coins(self) -> int:
         return self.__coin_observer.get_value()
 
-    def add_coins(self,value :int) -> None:
+    def add_coins(self, value: int) -> None:
         self.__coin_observer.update(value)
 
     def configure_level(
