+++ src/level/observers/__init__.py
@@ -1,0 +1,4 @@
+from .coin_observer import CoinObserver
+from .score_observer import ScoreObserver
+
+__all__ = ["ScoreObserver", "CoinObserver"]
