+++ src/scenes/concretes/level/level_scene_tick.py
@@ -4,8 +4,7 @@ from pygame import time
 
 from src.enums import GameEvent, HeroState, SceneAction
 from src.level import ILevelManager
-from src.utils import TO_SECONDS
-from src.utils import COIN_LIMIT
+from src.utils import COIN_LIMIT, TO_SECONDS
 
 from ...abstractions import Tick
 
@@ -39,10 +38,9 @@ class LevelSceneTick(Tick):
         hero.update(game_events, obstacles_manager.get_sprites(), camera)
         hero.animate()
 
-        if self.level_manager.get_coins() == COIN_LIMIT+1:
+        if self.level_manager.get_coins() == COIN_LIMIT + 1:
             self.level_manager.add_coins(1)
             self.level_manager.set_lives(self.level_manager.get_lives() + 1)
-
 
         if (
             hero.get_hero_state() == HeroState.DEAD
