+++ src/utils/constants.py
@@ -19,8 +19,6 @@ DEFAULT_R_TWO_SENSIBILITY = 0.6
 # UI CONSTANTS
 FONT_SIZE = 30
 
-PLAYER_LIVES = 3
-
 # VIDEO CONSTANTS
 BGR_FORMAT = "BGR"
 VIDEO_SCREEN_POSITION = (0, 0)
@@ -31,6 +29,7 @@ ANIMATION_INTERVAL = 175
 
 # PLAYER CONSTANTS
 HERO_SPEED = 6
+HERO_LIVES = 3
 MAX_GRAVITY = 10
 JUMP_VELOCITY = 30
 DEAD_FALL_THRESHOLD = 100
@@ -98,4 +97,6 @@ VALVOOPA_SIZE = (60, 90)
 
 # SCORE LIMIT
 SCORE_LIMIT = 999_999
+
+# COIN LIMIT
 COIN_LIMIT = 99
